Moving to renderer and using ts
